app/requirements.txt
uvicorn==0.30.6
fastapi==0.112.2
pydantic==2.6.4
python-dotenv==1.0.1
pytz==2023.3
SQLAlchemy==2.0.34
psycopg2==2.9.9
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
google-auth==2.20.0
requests==2.31.0
python-multipart==0.0.17
httpx==0.27.2